requirements.txt
langchain==0.0.201
python-dotenv==1.0.0
streamlit==1.23.1
openai==0.27.8
tiktoken==0.4.0
faiss-cpu==1.7.4
watchdog==3.0.0
unstructured==0.7.5
requests==2.31.0
scrapy==2.9.0